broke compiler util into 2 files
