+++ src/Compiler/CMakeLists.txt
@@ -181,6 +181,23 @@ add_onnx_mlir_library(OMCompilerUtils
   MC
   )
 
+  add_onnx_mlir_library(OMOptionUtils
+  OptionUtils.cpp
+
+  EXCLUDE_FROM_OM_LIBS
+
+  DEPENDS
+
+  INCLUDE_DIRS PRIVATE
+  ${FILE_GENERATE_DIR}
+
+  INCLUDE_DIRS PUBLIC
+  ${ONNX_MLIR_SRC_ROOT}/include
+
+  LINK_LIBS PUBLIC
+  )
+
+
 # CompilerUtils does not require cruntime or jniruntime to build,
 # however, they are required for execution when using the EmitLib or EmitJNI
 # options
